+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/RreadMessage.java
@@ -43,6 +43,9 @@ import uk.ac.rdg.resc.jstyx.StyxUtils;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:15  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -71,7 +74,7 @@ public class RreadMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public RreadMessage(int length, int type, int tag)
+    public RreadMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Rread";
@@ -79,7 +82,7 @@ public class RreadMessage extends StyxMessage
     
     public RreadMessage(int count, ByteBuffer data)
     {
-        this(0, 117, 0); // We'll set the length and tag later
+        this(0, (short)117, 0); // We'll set the length and tag later
         this.setData(count, data);
     }
