+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/TopenMessage.java
@@ -36,6 +36,9 @@ package uk.ac.rdg.resc.jstyx.messages;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:15  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -62,7 +65,7 @@ public class TopenMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public TopenMessage(int length, int type, int tag)
+    public TopenMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Topen";
@@ -73,7 +76,7 @@ public class TopenMessage extends StyxMessage
      */
     public TopenMessage(long fid, int mode)
     {
-        this(12, 112, 0); // The tag will be set later
+        this(12, (short)112, 0); // The tag will be set later
         this.fid = fid;
         this.mode = mode;
     }
