+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/TwriteMessage.java
@@ -43,6 +43,9 @@ import uk.ac.rdg.resc.jstyx.types.ULong;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:16  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -74,7 +77,7 @@ public class TwriteMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public TwriteMessage(int length, int type, int tag)
+    public TwriteMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Twrite";
@@ -82,7 +85,7 @@ public class TwriteMessage extends StyxMessage
     
     public TwriteMessage(long fid, ULong offset, int count, ByteBuffer data)
     {
-        this(0, 118, 0); // The tag and length will be set later
+        this(0, (short)118, 0); // The tag and length will be set later
         this.fid = fid;
         this.offset = offset;
         this.count = count;
